Created Notebook Store Service
Created initial notebook store service which handles/stores the data retrieved from the web api. This service will act like a 'mini' database for the notebook data.
Views should subscribe to the data being exposed from this store to centralize data shown on views.
The 'messages' will be used as the data being notified by the store. This is because the store will use the observer/subscriber pattern.